eni_scanner/requirements.txt
aiohttp
beautifulsoup4
paramiko
pysocks
pysnmp
aiosnmp
zeroconf
scapy
tqdm
pyyaml
cryptography
openpyxl
requests
